:ok_hand: F06 > ORDER > create NavBarRightSide.js
